fix: changed return type at services

+++ BAL/Services/OrderServices/OrderService.cs
@@ -1,4 +1,5 @@
-﻿using BAL.Dto.OrderDtos;
+﻿using BAL.Dto;
+using BAL.Dto.OrderDtos;
 using BAL.Dto.OrderlineDtos;
 using BAL.Mapper;
 using BAL.Services.OrderlineServices;
@@ -33,7 +34,7 @@ namespace BAL.Services.OrderServices
             _logger = logger;
         }
 
-        public async Task<Result<OrderDto>> CreateOrderAsync(OrderCreateDto request, CancellationToken cancellationToken)
+        public async Task<Result<string>> CreateOrderAsync(OrderCreateDto request, CancellationToken cancellationToken)
         {
             var validationResult = await _createValidator.ValidateAsync(request, cancellationToken);
 
@@ -42,31 +43,33 @@ namespace BAL.Services.OrderServices
                 return Result.Fail(validationResult.Errors.Select(failure => failure.ErrorMessage));
             }
 
-            var order = new Order
-            {
-                Description = request.Description,
-                CreatedAt = DateTime.UtcNow,
-            };
-
-            List<string>? failedOrderlineMessages = null;
+            var response = new CreateResponse();
 
             using var transaction = _orderRepository.BeginTransaction(IsolationLevel.Serializable);
 
             try
             {
+                var order = new Order
+                {
+                    Description = request.Description,
+                    CreatedAt = DateTime.UtcNow,
+                };
+
                 await _orderRepository.CreateOrderAsync(order, cancellationToken);
 
                 if (request.EquipmentToOrder is not null)
                 {
                     var orderlines = await _orderlineService.CreateOrderlinesAsync(request.EquipmentToOrder, order.Id, cancellationToken);
 
-                    failedOrderlineMessages = orderlines
-                        .Where(ol => ol.IsFailed)
-                        .SelectMany(ol => ol.Errors.Select(x => x.Message))
-                        .ToList();
+                    response.SuccessfulOrderlines = order.OrderLines.Count();
+                    response.TotalOrderlineRequestCount = request.EquipmentToOrder.Count();
+                    response.Errors = orderlines
+                        .Where(x => x.IsFailed)
+                        .SelectMany(x => x.Errors
+                        .Select(x => x.Message)).ToList();
                 }
 
-                order.Price = GetOrderPrice(order.OrderLines, cancellationToken);
+                order.Price = GetOrderPrice(order.OrderLines);
 
                 await _orderRepository.UpdateOrderAsync(order, cancellationToken);
 
@@ -79,37 +82,70 @@ namespace BAL.Services.OrderServices
                 transaction.Rollback();
             }
 
-            return order.MapToResponse(failedOrderlineMessages);
+            return response.GetResponseMessage();
         }
 
-        public async Task<Result> DeleteOrderAsync(Guid orderId, CancellationToken cancellationToken)
+        public async Task<Result<string>> DeleteOrderAsync(Guid orderId, CancellationToken cancellationToken)
         {
-            var order = await _orderRepository.GetOrderByIdAsync(orderId, cancellationToken);
+            using var transaction = _orderRepository.BeginTransaction(IsolationLevel.Serializable);
 
-            if(order is null)
+            try
             {
-                return Result.Fail(Errors.OrderDoesntExist);
+                var order = await _orderRepository.GetOrderByIdAsync(orderId, cancellationToken);
+
+                if (order is null)
+                {
+                    return Result.Fail(Errors.OrderDoesntExist);
+                }
+                
+                if(order.OrderLines is not null)
+                {
+                    await _orderlineService.DeleteOrderlineAsync(order.OrderLines, cancellationToken);
+                }
+
+                await _orderRepository.DeleteOrderAsync(order, cancellationToken);
+
+                transaction.Commit();
+
+                return DeleteResponse.DeleteSuccessfuly;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Error occured while deleting order {orderId}: {ErrorMessage}", orderId, ex.Message);
+                
+                transaction.Rollback();
+
+                return Result.Fail(DeleteResponse.DeleteFailed);
+            }
+        }
+
+        public async Task<Result<List<OrderDto>>> GetOrderWithPaginationAsync(int pageIndex, int pageSize, CancellationToken cancellationToken)
+        {
+            if(pageSize <= 0 || pageIndex <= 0)
+            {
+                return Result.Fail(Errors.PaginationParametersBelowZero);
             }
 
-            await _orderlineService.DeleteOrderlineAsync(order.OrderLines, cancellationToken);
-            
-            await _orderRepository.DeleteOrderAsync(order, cancellationToken);
+            var totalItems = await _orderRepository.GetTotalNumberOfOrders(cancellationToken);
+            var totalPages = (int)Math.Ceiling((double)totalItems / pageSize);
 
-            return Result.Ok();
-        }
+            if(totalPages < pageIndex)
+            {
+                return Result.Fail(Errors.EmptyPages(totalPages));
+            }
 
-        public async Task<Result<List<OrderDto>>> GetOrderWithPaginationAsync(CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
+            var orders = await _orderRepository.GetOrderPaginationByCreatedDateAsync(pageIndex, pageSize, cancellationToken);
+
+            return orders.MapToResponse();
         }
 
         public async Task<Result<OrderDto>> UpdateOrderAsync(Guid orderId, OrderUpdateDto updatedOrder, CancellationToken cancellationToken)
         {
-            var result = await _updateValidator.ValidateAsync(updatedOrder, cancellationToken);
+            var validationRezult = await _updateValidator.ValidateAsync(updatedOrder, cancellationToken);
 
-            if (!result.IsValid)
+            if (!validationRezult.IsValid)
             {
-                return Result.Fail(result.Errors.Select(failure => failure.ErrorMessage));
+                return Result.Fail(validationRezult.Errors.Select(failure => failure.ErrorMessage));
             }
 
             var order = await _orderRepository.GetOrderByIdAsync(orderId, cancellationToken);
@@ -119,16 +155,19 @@ namespace BAL.Services.OrderServices
                 return Result.Fail(Errors.OrderDoesntExist);
             }
 
-            //order.Description = updatedOrder.Description ?? order.Description;
-            //order.Price = updatedOrder.Price ?? order.Price;
-            //order.UpdatedAt = DateTime.Now;
+            if (updatedOrder.EquipmentToOrder is not null)
+            {
+                var result = await _orderlineService.UpdateOrderlineAsync(order, updatedOrder.EquipmentToOrder, cancellationToken);
+            }
+            
 
+            order.UpdatedAt = DateTime.UtcNow;
             await _orderRepository.UpdateOrderAsync(order, cancellationToken);
 
             return Result.Ok();
         }
 
-        public decimal GetOrderPrice(List<OrderLine>? orderLines, CancellationToken cancellationToken)
+        public decimal GetOrderPrice(List<OrderLine>? orderLines)
         {
             return orderLines is null ? 0 : orderLines.Sum(x => x.Price * x.Amount);
         }
